requirements.txt
arrow==1.2.3
async-timeout==4.0.2
coloredlogs==15.0.1
humanfriendly==10.0
paho-mqtt==1.6.1
pymodbus==2.5.3
pyserial==3.5
python-dateutil==2.8.2
python-snap7==1.3
redis==4.5.4
six==1.16.0
